seperated component from app file
